requirements.txt
absl-py==1.4.0
addict==2.4.0
albumentations==1.3.1
altgraph==0.17.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
efficientnet-pytorch==0.7.1
einops==0.6.1
et-xmlfile==1.1.0
filelock==3.12.2
fonttools==4.40.0
fsspec==2023.6.0
functional==0.1
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.1
jpegio==0.2.8
kiwisolver==1.4.4
lazy_loader==0.3
loguru==0.7.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.1
mmcv-full==1.7.0
mmengine==0.8.0
model-index==0.1.11
munch==4.0.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opendatalab==0.0.9
openmim==0.3.9
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.8.0
pretrainedmodels==0.7.4
prettytable==3.8.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.8
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
segmentation-models-pytorch==0.3.3
six==1.16.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.7.4
timm==0.9.2
tomli==2.0.1
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
yacs==0.1.8
yapf==0.40.1
zipp==3.15.0
